Move QualityOracle senders to 'core/quality' package
